backend/requirements.txt
# מחזיק את כל הספריות שהפרויקט צריך בשביל לרוץ

fastapi
uvicorn
sqlalchemy
psycopg2-binary
python-dotenv
pydantic[email]
bcrypt
python-jose[cryptography]
python-dotenv
apscheduler
pytz


